feat: Add Catalyst dark theme.
