+++ catalyst_voices/packages/catalyst_voices_assets/assets/colors/colors.xml
@@ -39,4 +39,43 @@
     <color name="light_on_surface_error_08">#14CC0000</color>
     <color name="light_on_surface_error_012">#1FCC0000</color>
     <color name="light_on_surface_error_016">#29CC0000</color>
+    <color name="dark_text_primary">#FFFFFF</color>
+    <color name="dark_text_on_primary">#E6E9F0</color>
+    <color name="dark_text_on_primary_container">#0C288D</color>
+    <color name="dark_text_disabled">#61D9DEE8</color>
+    <color name="dark_primary">#728EF3</color>
+    <color name="dark_on_primary">#0C288D</color>
+    <color name="dark_primary_container">#1035BC</color>
+    <color name="dark_on_primary_container">#E8ECFD</color>
+    <color name="dark_secondary">#DF8AF5</color>
+    <color name="dark_on_secondary">#26042F</color>
+    <color name="dark_secondary_container">#9910BC</color>
+    <color name="dark_on_secondary_container">#9910BC</color>
+    <color name="dark_outline">#7F90B3</color>
+    <color name="dark_outline_variant">#364463</color>
+    <color name="dark_error">#FF9999</color>
+    <color name="dark_on_error">#380000</color>
+    <color name="dark_error_container">#AD0000</color>
+    <color name="dark_on_error_container">#FFD1D1</color>
+    <color name="dark_success">#BAEDC2</color>
+    <color name="dark_on_success">#08210C</color>
+    <color name="dark_success_container">#1D722A</color>
+    <color name="dark_on_success_container">#CEF3D4</color>
+    <color name="dark_warning">#FBC9A7</color>
+    <color name="dark_on_warning">#2C1302</color>
+    <color name="dark_warning_container">#B64E07</color>
+    <color name="dark_on_warning_container">#FDE1CE</color>
+    <color name="dark_on_surface_neutral_08">#1FBFC8D9</color>
+    <color name="dark_on_surface_neutral_012">#1FBFC8D9</color>
+    <color name="dark_on_surface_neutral_016">#29212A3D</color>
+    <color name="dark_on_surface_primary_container">#0C288D</color>
+    <color name="dark_on_surface_primary_08">#1F123CD3</color>
+    <color name="dark_on_surface_primary_012">#1F123CD3</color>
+    <color name="dark_on_surface_primary_016">#29123CD3</color>
+    <color name="dark_on_surface_secondary_08">#1FD972F3</color>
+    <color name="dark_on_surface_secondary_012">#1FC014EB</color>
+    <color name="dark_on_surface_secondary_016">#29C014EB</color>
+    <color name="dark_on_surface_error_08">#1FFFC2C2</color>
+    <color name="dark_on_surface_error_012">#1FCC0000</color>
+    <color name="dark_on_surface_error_016">#29CC0000</color>
 </resources>
