+++ catalyst_voices/packages/catalyst_voices_brands/lib/src/themes/extended_color_scheme.dart
@@ -1,0 +1,179 @@
+import 'package:flutter/material.dart';
+
+@immutable
+class ExtendedColorScheme extends ThemeExtension<ExtendedColorScheme> {
+  const ExtendedColorScheme({
+    required this.textPrimary,
+    required this.textOnPrimary,
+    required this.textOnPrimaryContainer,
+    required this.textDisabled,
+    required this.success,
+    required this.onSuccess,
+    required this.successContainer,
+    required this.onSuccessContainer,
+    required this.warning,
+    required this.onWarning,
+    required this.warningContainer,
+    required this.onWarningContainer,
+    required this.onSurfaceNeutral08,
+    required this.onSurfaceNeutral012,
+    required this.onSurfaceNeutral016,
+    required this.onSurfacePrimaryContainer,
+    required this.onSurfacePrimary08,
+    required this.onSurfacePrimary012,
+    required this.onSurfacePrimary016,
+    required this.onSurfaceSecondary08,
+    required this.onSurfaceSecondary012,
+    required this.onSurfaceSecondary016,
+    required this.onSurfaceError08,
+    required this.onSurfaceError012,
+    required this.onSurfaceError016,
+  });
+
+  final Color? textPrimary;
+  final Color? textOnPrimary;
+  final Color? textOnPrimaryContainer;
+  final Color? textDisabled;
+  final Color? success;
+  final Color? onSuccess;
+  final Color? successContainer;
+  final Color? onSuccessContainer;
+  final Color? warning;
+  final Color? onWarning;
+  final Color? warningContainer;
+  final Color? onWarningContainer;
+  final Color? onSurfaceNeutral08;
+  final Color? onSurfaceNeutral012;
+  final Color? onSurfaceNeutral016;
+  final Color? onSurfacePrimaryContainer;
+  final Color? onSurfacePrimary08;
+  final Color? onSurfacePrimary012;
+  final Color? onSurfacePrimary016;
+  final Color? onSurfaceSecondary08;
+  final Color? onSurfaceSecondary012;
+  final Color? onSurfaceSecondary016;
+  final Color? onSurfaceError08;
+  final Color? onSurfaceError012;
+  final Color? onSurfaceError016;
+
+  @override
+  ThemeExtension<ExtendedColorScheme> copyWith({
+    Color? textPrimary,
+    Color? textOnPrimary,
+    Color? textOnPrimaryContainer,
+    Color? textDisabled,
+    Color? success,
+    Color? onSuccess,
+    Color? successContainer,
+    Color? onSuccessContainer,
+    Color? warning,
+    Color? onWarning,
+    Color? warningContainer,
+    Color? onWarningContainer,
+    Color? onSurfaceNeutral08,
+    Color? onSurfaceNeutral012,
+    Color? onSurfaceNeutral016,
+    Color? onSurfacePrimaryContainer,
+    Color? onSurfacePrimary08,
+    Color? onSurfacePrimary012,
+    Color? onSurfacePrimary016,
+    Color? onSurfaceSecondary08,
+    Color? onSurfaceSecondary012,
+    Color? onSurfaceSecondary016,
+    Color? onSurfaceError08,
+    Color? onSurfaceError012,
+    Color? onSurfaceError016,
+  }) {
+    return ExtendedColorScheme(
+      textPrimary: textPrimary ?? this.textPrimary,
+      textOnPrimary: textOnPrimary ?? this.textOnPrimary,
+      textOnPrimaryContainer:
+          textOnPrimaryContainer ?? this.textOnPrimaryContainer,
+      textDisabled: textDisabled ?? this.textDisabled,
+      success: success ?? this.success,
+      onSuccess: onSuccess ?? this.onSuccess,
+      successContainer: successContainer ?? this.successContainer,
+      onSuccessContainer: onSuccessContainer ?? this.onSuccessContainer,
+      warning: warning ?? this.warning,
+      onWarning: onWarning ?? this.onWarning,
+      warningContainer: warningContainer ?? this.warningContainer,
+      onWarningContainer: onWarningContainer ?? this.onWarningContainer,
+      onSurfaceNeutral08: onSurfaceNeutral08 ?? this.onSurfaceError08,
+      onSurfaceNeutral012: onSurfaceNeutral012 ?? this.onSurfaceError012,
+      onSurfaceNeutral016: onSurfaceNeutral016 ?? this.onSurfaceError016,
+      onSurfacePrimaryContainer:
+          onSurfacePrimaryContainer ?? this.onSurfacePrimaryContainer,
+      onSurfacePrimary08: onSurfacePrimary08 ?? this.onSurfacePrimary08,
+      onSurfacePrimary012: onSurfacePrimary012 ?? this.onSurfacePrimary012,
+      onSurfacePrimary016: onSurfacePrimary016 ?? this.onSurfacePrimary016,
+      onSurfaceSecondary08: onSurfaceSecondary08 ?? this.onSurfaceSecondary08,
+      onSurfaceSecondary012:
+          onSurfaceSecondary012 ?? this.onSurfaceSecondary012,
+      onSurfaceSecondary016:
+          onSurfaceSecondary016 ?? this.onSurfaceSecondary016,
+      onSurfaceError08: onSurfaceError08 ?? this.onSurfaceError08,
+      onSurfaceError012: onSurfaceError012 ?? this.onSurfaceError012,
+      onSurfaceError016: onSurfaceError016 ?? this.onSurfaceError016,
+    );
+  }
+
+  @override
+  ExtendedColorScheme lerp(
+    ThemeExtension<ExtendedColorScheme>? other,
+    double t,
+  ) {
+    if (other is! ExtendedColorScheme) {
+      return this;
+    }
+    return ExtendedColorScheme(
+      textPrimary: Color.lerp(textPrimary, other.textPrimary, t),
+      textOnPrimary: Color.lerp(textOnPrimary, other.textOnPrimary, t),
+      textOnPrimaryContainer: 
+        Color.lerp(textOnPrimaryContainer, other.textOnPrimaryContainer, t),
+      textDisabled: Color.lerp(textDisabled, other.textDisabled, t),
+      success: Color.lerp(success, other.success, t),
+      onSuccess: Color.lerp(onSuccess, other.onSuccess, t),
+      successContainer: Color.lerp(successContainer, other.successContainer, t),
+      onSuccessContainer:
+        Color.lerp(onSuccessContainer, other.onSuccessContainer, t),
+      warning: Color.lerp(warning, other.warning, t),
+      onWarning: Color.lerp(onWarning, other.onWarning, t),
+      warningContainer: Color.lerp(warningContainer, other.warningContainer, t),
+      onWarningContainer:
+        Color.lerp(onWarningContainer, other.onWarningContainer, t),
+      onSurfaceNeutral08: 
+        Color.lerp(onSurfaceNeutral08, other.onSurfaceNeutral08, t),
+      onSurfaceNeutral012:
+        Color.lerp(onSurfaceNeutral012, other.onSurfaceNeutral012, t),
+      onSurfaceNeutral016:
+        Color.lerp(onSurfaceNeutral016, other.onSurfaceNeutral016, t),
+      onSurfacePrimaryContainer: Color.lerp(
+        onSurfacePrimaryContainer,
+        other.onSurfacePrimaryContainer,
+        t,
+      ),
+      onSurfacePrimary08: 
+        Color.lerp(onSurfacePrimary08, other.onSurfacePrimary08, t),
+      onSurfacePrimary012: 
+        Color.lerp(onSurfacePrimary012, other.onSurfacePrimary012, t),
+      onSurfacePrimary016:
+          Color.lerp(onSurfacePrimary016, other.onSurfacePrimary016, t),
+      onSurfaceSecondary08:
+          Color.lerp(onSurfaceSecondary08, other.onSurfaceSecondary08, t),
+      onSurfaceSecondary012:
+          Color.lerp(onSurfaceSecondary012, other.onSurfaceSecondary012, t),
+      onSurfaceSecondary016:
+          Color.lerp(onSurfaceSecondary016, other.onSurfaceSecondary016, t),
+      onSurfaceError08: Color.lerp(onSurfaceError08, other.onSurfaceError08, t),
+      onSurfaceError012:
+          Color.lerp(onSurfaceError012, other.onSurfaceError012, t),
+      onSurfaceError016:
+          Color.lerp(onSurfaceError016, other.onSurfaceError016, t),
+    );
+  }
+}
+
+extension ExtendedColorsExtension on ThemeData {
+  ExtendedColorScheme get colors => extension<ExtendedColorScheme>()!;
+  Color get linksPrimary => primaryColor;
+}
