+++ catalyst_voices/packages/catalyst_voices_brands/lib/src/themes/fallback.dart
@@ -2,4 +2,6 @@ import 'package:flutter/material.dart';
 
 /// [ThemeData] for the `fallback` brand.
 final ThemeData fallback = ThemeData.light();
+
+/// Dark [ThemeData] for the `fallback` brand.
 final ThemeData darkFallback = ThemeData.dark();
